Added calc_distance function and updated init_board function
Added calc_distance(x1, y1, x2, y2) where x1 and y1 are the xy coordinates of point 1, while x2 and y2 are the xy coordinates of point 2. It will return the distance between these two points, which also happen to be the least number of turns it takes to get from point 1 to point 2, assuming nothing is cutting them off. Also updated(init_board) so that when filling boardtiles taken by snakes, it fills it with the snake's ID rather than just "snake". A board tile has a 0 if empty, and "food" if it contains food.